autopilot/requirements.txt
tweepy
requests
crewai
langchain-google-genai
python-dotenv
python-frontmatter
google-generativeai